+++ src/main/java/kr/ync/domain/MessageVO.java
@@ -5,7 +5,7 @@ import lombok.Data;
 @Data
 public class MessageVO {
 	
-	private int message_idx;
+	private Long message_idx;
 	private String id;
 	private String content;
 	private String receivce_id;
